Move C++ ImportType to its own file
