+++ crates/c-api/tests/CMakeLists.txt
@@ -23,7 +23,8 @@ add_capi_test(types FILES
   val_type.cc
   global_type.cc
   table_type.cc
-  func_type.cc)
+  func_type.cc
+  import_type.cc)
 add_capi_test(func FILES func.cc)
 add_capi_test(component-instantiate FILES component/instantiate.cc)
 add_capi_test(error FILES error.cc)
